# swagger old.yaml
openapi: '3.0.0'
info:
  title: Wheelancer API with Swagger
  version: '1.0'
  description: "This is a simple CRUD API application made with Express and documented with Swagger 
  <br><br><h1>AUTH PASS (🔑): admin
  <br><br><h1>SUPER_AUTH PASS (🔑🦹‍♂️): superadmin for System API calls and Admin registeration."
servers:
  - url: http://localhost:8080/
  


#---<<<<<<<<<<<<<<<<<<<<<<<<<<<USER>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
components:
...
  securitySchemes:
    Authorization:          
      type: apiKey
      name: Authorization
      in: header
      description: 'Use that token => admin'
    Super-Authorization:          
      type: apiKey
      name: super_auth
      in: header
      description: 'Super Special key for creating couriers => superadmin'  
security:
   - Authorization: []    
paths:
...
  /system/inserttuples:
    post:
      security:
        - Super-Authorization: []
      tags:
        - "System"

      description: Insert example tuples.
      responses:
        '200':
          description: OK - Inserted
        '500':
          description: MYSQL ERROR.
  /system/truncatetables:
    delete:
      security:
        - Super-Authorization: []
      tags:
        - "System"

      description: Truncate tables.
      responses:
        '200':
          description: OK - Truncated.
        '500':
          description: MYSQL ERROR.
  /system/inserttables:
    post:
      security:
        - Super-Authorization: []
      tags:
        - "System"

      description: <h1>Creates tables (Do not use until fully implemented)🧰.
      responses:
        '200':
          description: OK - Inserted.
        '500':
          description: MYSQL ERROR.



#---<<<<<<<<<<<<<<<<<<<<<<<<<<<Forum>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
...